+++ scripts/utils/lighting.js
@@ -17,6 +17,18 @@ export class Lighting {
     }
   }
 
+  static setLightLevel(darknessLevel) {
+    if (darknessLevel >= 0 && darknessLevel < 0.5) {
+      return 2;
+    }
+    if (darknessLevel >= 0.5 && darknessLevel < 0.75) {
+      return 1;
+    }
+    if (darknessLevel >= 0.75 && darknessLevel <= 1) {
+      return 0;
+    }
+  }
+
   static async show_lightLevel_box(selected_token, tokenHUD, html) {
     if (selected_token.actor.system.attributes.hp.value > 0) {
       // Determine lightLevel of the token (dark,dim,light)
@@ -52,22 +64,46 @@ export class Lighting {
 
   static async find_token_lighting(selected_token) {
     let lightLevel = 0;
+
+    // placed drawings with light overrides (perfect-vision)
+    if (canvas.drawings.placeables) {
+      for (const placed_drawing of canvas.drawings.placeables) {
+        if (placed_drawing.document.flags['perfect-vision'].enabled) {
+          let result = Core.placedDrawingsContain(placed_drawing.document,selected_token);
+          Core.log("result:",selected_token.actor.name,result);
+
+          if (result) {
+            let drawingOverride = placed_drawing.document.flags['perfect-vision'].darkness;
+            let drawingLightLevel = this.setLightLevel(drawingOverride);
+            Core.log("return lightLevel:",drawingLightLevel);
+            if (drawingLightLevel > lightLevel) {
+              lightLevel = drawingLightLevel;
+              Core.log("set lightLevel:",lightLevel);
+            }
+          }
+        }
+        else {
+          Core.log("no darkness:",placed_drawing);
+        }
+      }
+    }
+
     // placed lights
     if (canvas.lighting.objects) {
-      for (const placed_lights of canvas.lighting.objects.children) {
-        if (placed_lights.source.active == true) {
-          let tokenDistance = this.get_calculated_light_distance(selected_token, placed_lights);
-          let foundWall = Core.get_wall_collision(selected_token, placed_lights);
+      for (const placed_light of canvas.lighting.objects.children) {
+        if (placed_light.source.active == true) {
+          let tokenDistance = this.get_calculated_light_distance(selected_token, placed_light);
+          let foundWall = Core.get_wall_collision(selected_token, placed_light);
           if (!foundWall) {
             // check for dim
-            if (tokenDistance <= placed_lights.document.config.dim) {
-              if ((lightLevel < 1) && (placed_lights.document.config.dim > 0)) {
+            if (tokenDistance <= placed_light.document.config.dim) {
+              if ((lightLevel < 1) && (placed_light.document.config.dim > 0)) {
                 lightLevel = 1;
               }
             }
             // check for bright
-            if (tokenDistance <= placed_lights.document.config.bright) {
-              if ((lightLevel < 2) && (placed_lights.document.config.bright > 0)) {
+            if (tokenDistance <= placed_light.document.config.bright) {
+              if ((lightLevel < 2) && (placed_light.document.config.bright > 0)) {
                 lightLevel = 2;
               }
             }
@@ -77,21 +113,21 @@ export class Lighting {
     }
     // placed tokens
     if (canvas.tokens.placeables) {
-      for (const placed_tokens of canvas.tokens.placeables) {
-        if (placed_tokens.actor) {
-          if (placed_tokens.light.active == true) {
-            let tokenDistance = Core.get_calculated_distance(selected_token, placed_tokens);
-            let foundWall = Core.get_wall_collision(selected_token, placed_tokens);
+      for (const placed_token of canvas.tokens.placeables) {
+        if (placed_token.actor) {
+          if (placed_token.light.active == true) {
+            let tokenDistance = Core.get_calculated_distance(selected_token, placed_token);
+            let foundWall = Core.get_wall_collision(selected_token, placed_token);
             if (!foundWall) {
               // check for dim
-              if (tokenDistance <= placed_tokens.document.light.dim) {
-                if ((lightLevel < 1) && (placed_tokens.document.light.dim > 0)) {
+              if (tokenDistance <= placed_token.document.light.dim) {
+                if ((lightLevel < 1) && (placed_token.document.light.dim > 0)) {
                   lightLevel = 1;
                 }
               }
               // check for bright
-              if (tokenDistance <= placed_tokens.document.light.bright) {
-                if ((lightLevel < 2) && (placed_tokens.document.light.bright > 0)) {
+              if (tokenDistance <= placed_token.document.light.bright) {
+                if ((lightLevel < 2) && (placed_token.document.light.bright > 0)) {
                   lightLevel = 2;
                 }
               }
@@ -137,12 +173,46 @@ export class Lighting {
     let elevated_distance = 0;
     let gridSize = canvas.grid.size;
     let gridDistance = canvas.scene.grid.distance;
+    let z1Actual = 0;
+    let z2Actual = 0;
 
+    const x1 = selected_token.center.x;
+    const y1 = selected_token.center.y;
+    let z1 = selected_token.document.elevation;
+
+    const x2 = placed_lights.center.x;
+    const y2 = placed_lights.center.y;
+    let z2 = 0;
+
+    // If using Mod that adds elevation to light placements ( Levels for example )
+    if (placed_lights.document.elevation) {
+      z2 = placed_lights.document.elevation;
+     
+      if (placed_lights.t || placed_lights.b) {
+        let t = placed_lights.t;
+        let b = placed_lights.b;
+
+        if (t == 'Infinity') {t = z1;}
+        if (b == 'Infinity') {b = z1;}
+
+        if (z1 <= b) {z2 = b;}
+        if (z1 >= t) {z2 = t;}
+        // between top and bottom so, treat it as a pill
+        // treat the light as being at the same level as the token for calculation
+        if ((z1 >= t) && (z1 <= t)) {
+          z2 = z1;
+        }
+      }
+    }
+
+    // convert elevation (ft) to canvas grid values
+    z1Actual =  (z1 / gridDistance) * gridSize;
+    z2Actual = (z2 / gridDistance) * gridSize;
+    
     // Measure grid distance with elevation
-    let e1 = Math.abs((selected_token.center.x - placed_lights.document.x));
-    let e2 = Math.abs((selected_token.center.y - placed_lights.document.y));
-    // lights don't have elevation?
-    let e3 = Math.abs(((selected_token.document.elevation/gridDistance)* gridSize));
+    let e1 = Math.abs(x1 - x2);
+    let e2 = Math.abs(y1 - y2);
+    let e3 = Math.abs(z1Actual - z2);
     let distance = Math.sqrt(e1*e1 + e2*e2 + e3*e3);
 
     elevated_distance = (distance / gridSize) * gridDistance;;